feature: move contextual commands to top level
Previously, a user would need to to specify a subcommand i.e. `budgey
pile add 10` to add 10 to a pile. But the pile specifier is not
necessary for a user. They might have changed piles and now just want to
add 10.

This changes this behaviour to not need the additional pile specifier ->
so now the command would become `budgey add 10`. This is vastly simpler
for the user.

Notes: I'm not happy with the implementation detail of this. We
duplicate the context call across ALL methods that aren't init. Future
efforts should find a way to structure the cli so we only do this if the
user did NOT do `budgey init` (or some future branch that doesn't
require context)