fix/refactor: separate top level commands out
Previously, all the top level commands were in budgey cli, but this
meant needing to get the context for all that weren't init.

This change separates out the cli into the init command, for creating
dependencies for context, and the `commands`.

This means we only need to get the context if the variant is `commands`,
reducing the need to duplicate context everywhere.