fix: remove being able to set initial balance
This causes some conflicts with how a source control would want to
handle this. We don't want some initial balance coming and potentially
messing up a history. If someone wants to add to a pile, they must
explicitly perform an add commit for now.

We can review making 'initial balance' more of a shorthand for creating
an add or withdraw commit in the future.