feat: implement chain subcommand to replace -t arg
This used to be `budgey pile -t` to see transactions but this was a
confusing api.

-t could be invoked after a pile subcommand, but didn't have an effect
on the output of a specific pile command. i.e. `budgey pile -t add 10`
would show a transaction chain after, but this format goes against
convention of having args before commands.

Now, a user will invoke this behaviour with `budgey chain` to see the
transaction chain. This is cleaner and avoids all the complications.